+++ platform/viewer/public/config/default.js
@@ -1,7 +1,6 @@
 window.config = {
   // default: '/'
   routerBasename: '/',
-  whiteLabelling: {},
   extensions: [],
   showStudyList: true,
   filterQueryParam: false,